Move activation and menu initialization to ApplicationDelegate
